Moved webpack to enable use for both demo and taxrules - again
