+++ packages/playground/src/demo/webapp/app/LoadComponent.tsx
