+++ packages/playground/src/demo/webapp/styles/react-tabs.scss
