+++ packages/playground/src/demo/webapp/styles/toolbarstyle.scss
